small fix & version bump

+++ pysmtb/image.py
@@ -63,7 +63,7 @@ def annotate_image(image, label, font_path=None, font_size=16, font_color=[1.], 
         draw.text((x, y), text=label, fill=tuple(font_color) + (255,), font=font, stroke_width=stroke_width, stroke_fill=tuple(stroke_color) + (255,))
     else:
         draw.text((x, y), text=label, fill=tuple(font_color) + (255,), font=font, stroke_width=stroke_width, stroke_fill=tuple(stroke_color) + (255,))
-    mask = np.atleast_3d(np.array(mask, dtype=np.float) / 255.).astype(image.dtype)
+    mask = np.atleast_3d(np.array(mask, dtype=np.float32) / 255.).astype(image.dtype)
     alpha = np.atleast_3d(mask[:,:,-1])
     mask = np.atleast_3d(mask[:,:,:-1])
